PartialRespawn(Phase2)
The CorrManager script now respawns the player in, and the corruption UI works as intended!

A respawnLoc empty is necessary as a child of PlayerList for this to work, and the player must be a child of respawnLoc. In addition, a script has been added to the cinemachine virtual cam (which is now a prefab) to find the player once they respawn.

Currently respawnLoc doesn't update location on talking with a terminal, but that's an easy fix later.