CrateCreator
Okay, there's now a script for creating crates at the push of a button. So I goofed up when I was with you, the script for the button is in a child object, I just didn't notice. Just keep it to what the prefab is.
Link the flag script to the cratespawner object I added in as a prefab. It should be good to go from there. Make sure "Hold" is toggled in the button script, which is in the child object "Button" on the PhysButton. GravMult seems to work well enough for me at 4.